refactor: 이벤트 등록하는 함수 분리
